EES-5275 Move query requests into their own file
