+++ src/GovUk.Education.ExploreEducationStatistics.Data.Api/Services/PermalinkService.cs
@@ -12,6 +12,7 @@ using GovUk.Education.ExploreEducationStatistics.Common;
 using GovUk.Education.ExploreEducationStatistics.Common.Extensions;
 using GovUk.Education.ExploreEducationStatistics.Common.Model;
 using GovUk.Education.ExploreEducationStatistics.Common.Model.Data.Query;
+using GovUk.Education.ExploreEducationStatistics.Common.Requests;
 using GovUk.Education.ExploreEducationStatistics.Common.Services.Interfaces;
 using GovUk.Education.ExploreEducationStatistics.Common.Utils;
 using GovUk.Education.ExploreEducationStatistics.Content.Model;
@@ -87,11 +88,11 @@ namespace GovUk.Education.ExploreEducationStatistics.Data.Api.Services
         public async Task<Either<ActionResult, PermalinkViewModel>> CreatePermalink(PermalinkCreateRequest request,
             CancellationToken cancellationToken = default)
         {
-            return await (request.ReleaseId ?? await FindLatestPublishedReleaseVersionId(request.QueryRequest.SubjectId))
+            return await (request.ReleaseId ?? await FindLatestPublishedReleaseVersionId(request.Query.SubjectId))
                 .OnSuccess(releaseVersionId =>
                 {
                     return _tableBuilderService.Query(releaseVersionId,
-                            request.QueryRequest.AsObservationQueryContext(),
+                            request.Query.AsObservationQueryContext(),
                             cancellationToken)
                         .OnSuccess<ActionResult, TableBuilderResultViewModel, PermalinkViewModel>(async tableResult =>
                         {
@@ -102,7 +103,7 @@ namespace GovUk.Education.ExploreEducationStatistics.Data.Api.Services
                             );
 
                             var csvMetaTask = _permalinkCsvMetaService.GetCsvMeta(
-                                request.QueryRequest.SubjectId,
+                                request.Query.SubjectId,
                                 tableResult.SubjectMeta,
                                 cancellationToken
                             );
@@ -137,7 +138,7 @@ namespace GovUk.Education.ExploreEducationStatistics.Data.Api.Services
                             var permalink = new Permalink
                             {
                                 ReleaseVersionId = releaseVersionId,
-                                SubjectId = request.QueryRequest.SubjectId,
+                                SubjectId = request.Query.SubjectId,
                                 PublicationTitle = subjectMeta.PublicationName,
                                 DataSetTitle = subjectMeta.SubjectName,
                             };
